market_analysis/requirements.txt
# Core dependencies for Market Trend Analysis System
networkx>=3.0
requests>=2.31.0
pinecone>=5.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Optional dependencies for extended functionality
# Uncomment if needed for specific use cases
# pillow>=10.0.0  # For image processing
# opencv-python>=4.8.0  # For video processing
# transformers>=4.30.0  # For advanced NLP if not using Grok API
